backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
pyttsx3==2.90
python-dotenv==1.0.0
google-generativeai==0.3.2
pypdf==5.1.0
